feat: Compile if more cases;
Working on adding support for all cases, this part is going to take some
more time. I'm also planning on adding multiple cases like if some
|| other || ... {};

I also wan't to support comparing both local and global variables, this
is going to take a bit more time and be a bit harder tho.